min: time seconds type propery

+++ model/struct.go
@@ -17,5 +17,5 @@ type Message struct {
 }
 
 type Time struct {
-	Seconds uint16
+	Seconds uint
 }
